Task 2 Completed
Task description: "The goal of the predict color proba method is, given a list of Color values (i.e. one value per avacado we are trying to predict), we would like to output the entire distribution over the GoodToEat discrete random variable. For every Color value, we want to return a list of tuples: one tuple per value that the GoodToEat attribute can take on. Inside of this tuple we want to store two things: the value of the GoodToEat attribute G (the first value in the tuple), and the probability Pr[GoodToEat = G|Color = c] (the second value in the tuple). Remember to calculate this value with Baye’s rule (the equation is above in the previous task). I recommend calculating each pmf at a time (i.e. processing one Color value at a time), calculating the numerator of Baye’s rule for the entire pmf, and then normalizing the pmf by the total of the pmf. Please collect these pmfs in a list (with each pmf at the same index as the corresponding Color value that generated it in the input list) and return this list."