Finished Extra Credit
Please add two methods to your AvacadoPredictor class: predict color and predict softness. Again, let us only talk about predict color and the same process will be true for predict softness (of course focusing on Softness rather than Color). The goal of this is to actually produce an prediction for each input example. predict color will behave similarly to predict color proba, the only difference is that instead of producing an entire pmf: we want to produce the mostly likely value of the GoodToEat attribute. I recommend calling predict color proba within predict color and processing each pmf to find the mostly probably value of GoodToEat. Your output should be a list of GoodToEat enum values that correspond to the predictions that we want to make for each of the input Colors.