Finished Task 1.1, Task 1.2 needs fixing
Task 1: class BinomialDistribution (50 points)
A Binomial distribution has two parameters: n and p, where p is the probability of a single trial being in the “success” state, and n is the number of trials. A Binomial distribution models the number of successes that occur in those n trials. In your fit method, be sure to calculate p using the MLE estimate discussed in lecture, and be sure to use it in your prob method!