Finished task 2.1 (fit/lamda)
Task 2: class PoissonDistribution (50 points)
A Binomial distribution has one parameter: λ, where λ is the rate of (fair) trials occuring. A Poisson distribution models how long it will take for the first successful trial to occur. In your fit method, be sure to calculate λ using the MLE estimate discussed in lecture, and be sure to use it in your prob method!