Finished 3.1 **Extra Credit
A Gaussian distribution has two parameters: μ, and σ2, where μ is the center of the distribution, while σ2 is the spread. A Gaussian distribution models how likely an example is to occur given how far away it is from the center of the distribution according to the spread. In your fit method, be sure to calculate μ and σ2 using the MLE estimate discussed in lecture, and be sure to use them in your prob method!