Finished Task 2 and 3
Task 2 doesnt pass the autograder yet so will have to fix:

Once we have a way of making a random projection, we want to know if the random projection preserves the geometry of the point cloud. In this function you should check that the johnson- lindenstrauss contraints are met:
∀⃗xi,⃗xj (1−ε)||⃗xi −⃗xj||2 ≤||f(⃗xi)−f(⃗xj)||2 ≤(1+ε)||⃗xi −⃗xj||2
If a pair of points breaks a constraint, your function should return False, and True if all constraints are met.

task 3:

This function is the las vegas version. randomly project will do the projection, but the projection may not preserve the geometry of the point cloud (i.e. the output can be wrong). The way we will make this algorithm is as follows:
1. randomly project the data into k dimensions
2. check if the projection satisfies the constraints
3. repeat this process until a projection that satisfies the constraints is produced.
Your function should return three things (in the following order): the value of f that produced a valid projection, the valid projection itself, and the number of iterations it took to produce the valid projection.