Task 1/2/3 Completed
task 1/2 were basic setting:

task 3:

Task 3: def reconstruct X using k features (20 points)
This function should calculate the SVD of X and use the SVD of X to reconstruct X using only the best k features from the SVD. However, unlike the 99% use case of SVD where we want to create a matrix with k columns, this time we want to express the reconstruction in the original feature space (just using only the information contained within the top k new features). Here is how we do this:
• CalculateU,Σ,VT =svd(X)
• Set all singular values outside of the top k to be zero. • Reconstruct Xˆ = UΣV T
The subtlety is in step 2. Normally we would truncate U and Σ by throwing away parts of the matrices that we don’t want to keep. Here however we not truncating, but instead zeroing out singular values. Whenever we set a singular value to zero, we are effectively throwing it away when we do the matrix multiplication (i.e. those new features will have a coefficient of 0 in the matmul and will not contribute anything). This process however produces a matrix that has the same dimensionality as the original dataset, albeit containing less information than the original dataset.