task 6 Completed
Given a numpy array of z-scores, we want to figure out which of them are anomalous values. Using the value you set for ZSCORE STDDEV THRESHOLD earlier, figure out which indices in this array contain values that are anomalous (according to your threshold). Your output should also be a numpy array.
Note: anomalies can occur on either side of the mean!