Finished task 1 (theres only one task)
Task 1: class LinearRegression (100 points)
In the file linearRegression.py, you will find a class called LinearRegression. Remember, linear
regression learns a hyperplane (i.e. a line in higher dimensions) using the equation:
⃗θ∗ = [X,⃗1]†⃗ygt
where A† is the pseudoinverse of matrix A, ⃗ygt is the vector of ground truth values, and [X,⃗1] is
concatenating a column of 1s to feature matrix X.
When it is time to make predictions, we use the following equation:
ˆ⃗y = [X,⃗1]⃗θ∗
. Your job is complete the following methods:
• the fit method: This method is where you are handed training data X and ⃗ygt and need to
compute ⃗θ∗
• the predict method: This method is where you are handed feature matrix X and are asked to
make predictions on that data using the parameters ⃗θ∗ that you computed from the fit.